--- kernel/src/FrameBuffer.hpp
@@ -1,12 +1,0 @@
-#pragma once
-
-#include <stddef.h>
-
-struct Framebuffer
-{
-	void* BaseAddress;
-	size_t BufferSize;
-	unsigned int Width;
-	unsigned int Height;
-	unsigned int PixelsPerScanLine;
-};
