+++ kernel/src/Rendering/FrameBuffer.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <stddef.h>
+
+struct Framebuffer
+{
+	void* BaseAddress;
+	size_t BufferSize;
+	unsigned int Width;
+	unsigned int Height;
+	unsigned int PixelsPerScanLine;
+};
