+++ kernel/src/Types/Bitmap.hpp
@@ -1,0 +1,14 @@
+#pragma once
+#include <stddef.h>
+#include <stdint.h>
+
+class Bitmap
+{
+private:
+public:
+	size_t Size;
+	uint8_t* Buffer;
+
+	bool operator[]( uint64_t index );
+	void Set( uint64_t index, bool val );
+};
